+++ R/complete.R
@@ -6,6 +6,10 @@
 #' @param data A Polars Data/LazyFrame
 #' @param ... Any expression accepted by `dplyr::select()`: variable names,
 #'  column numbers, select helpers, etc.
+#'
+#' When used with continuous variables, you may need to fill in values that do
+#' not appear in the data: to do so use expressions like `year = 2010:2020` or
+#' `year = full_seq(year, 1)`.
 #' @param fill A named list that for each variable supplies a single value to
 #' use instead of `NA` for missing combinations.
 #' @param explicit Should both implicit (newly created) and explicit
@@ -52,11 +56,41 @@ complete.RPolarsDataFrame <- function(
   fill = list(),
   explicit = TRUE
 ) {
-  # if (isFALSE(explicit)) {
-  #   abort("`explicit = FALSE` is not supported by tidypolars.")
-  # }
-  vars <- tidyselect_dots(data, ...)
-  if (length(vars) < 2) {
+  # Determine the dots names for the relocate() call at the end. Seems like
+  # the simplest way to get this info and shouldn't be too expensive.
+  all_dots <- names(enquos(..., .named = TRUE))
+
+  # I want the separation named/unnamed dots, which is why I don't reuse the
+  # `all_dots` variable from above.
+  dots <- enquos(...)
+  names_dots <- names(dots)
+
+  unnamed_dots <- dots[which(names_dots == "")]
+  named_dots <- dots[which(names_dots != "")]
+
+  data_eval_dots <- build_data_context(data)
+  unnamed_dots <- lapply(unnamed_dots, function(x) {
+    tidyselect::eval_select(
+      x,
+      data_eval_dots,
+      error_call = caller_env()
+    )
+  }) |>
+    unlist() |>
+    names()
+
+  named_dots <- lapply(names(named_dots), function(x) {
+    out <- tibble(!!named_dots[[x]])
+    names(out) <- x
+    if (is_polars_df(data)) {
+      as_polars_df(out)
+    } else if (is_polars_lf(data)) {
+      as_polars_lf(out)
+    }
+  }) |>
+    set_names(names(named_dots))
+
+  if (length(unnamed_dots) < 2 && length(named_dots) == 0) {
     return(data)
   }
 
@@ -64,16 +98,34 @@ complete.RPolarsDataFrame <- function(
   mo <- attributes(data)$maintain_grp_order
   is_grouped <- !is.null(grps)
 
-  if (isTRUE(is_grouped)) {
-    chain <- data$group_by(grps, maintain_order = mo)$agg(
-      pl$col(vars)$unique()$sort()
-    )
+  chain_is_empty <- FALSE
+  if (length(unnamed_dots) == 0) {
+    chain <- if (is_polars_df(data)) {
+      pl$DataFrame()
+    } else if (is_polars_lf(data)) {
+      pl$LazyFrame()
+    }
+    chain_is_empty <- TRUE
   } else {
-    chain <- data$select(pl$col(vars)$unique()$sort()$implode())
+    if (isTRUE(is_grouped)) {
+      chain <- data$group_by(grps, maintain_order = mo)$agg(
+        pl$col(unnamed_dots)$unique()$sort()
+      )
+    } else {
+      chain <- data$select(pl$col(unnamed_dots)$unique()$sort()$implode())
+    }
   }
 
-  for (i in seq_along(vars)) {
-    chain <- chain$explode(vars[i])
+  for (i in seq_along(unnamed_dots)) {
+    chain <- chain$explode(unnamed_dots[i])
+  }
+
+  for (i in seq_along(named_dots)) {
+    if (chain_is_empty) {
+      chain <- named_dots[[i]]
+    } else {
+      chain <- chain$join(named_dots[[i]], how = "cross", join_nulls = TRUE)
+    }
   }
 
   # To avoid filling explicit missings, i.e. missings that were already in the
@@ -83,33 +135,36 @@ complete.RPolarsDataFrame <- function(
     if (isTRUE(is_grouped)) {
       already_exist <- chain$join(
         data,
-        on = c(grps, vars),
+        on = c(grps, all_dots),
         how = "inner",
         join_nulls = TRUE
       )
       dont_already_exist <- chain$join(
         data,
-        on = c(grps, vars),
+        on = c(grps, all_dots),
         how = "anti",
         join_nulls = TRUE
       )
     } else {
       already_exist <- chain$join(
         data,
-        on = vars,
+        on = all_dots,
         how = "inner",
         join_nulls = TRUE
       )
       dont_already_exist <- chain$join(
         data,
-        on = vars,
+        on = all_dots,
         how = "anti",
         join_nulls = TRUE
       )
     }
-    other_vars <- setdiff(names(already_exist), names(dont_already_exist))
+    other_unnamed_dots <- setdiff(
+      names(already_exist),
+      names(dont_already_exist)
+    )
     schema <- already_exist$schema
-    for (i in other_vars) {
+    for (i in other_unnamed_dots) {
       dont_already_exist <- dont_already_exist$with_columns(
         pl$lit(NA)$cast(schema[[i]])$alias(i)
       )
@@ -118,21 +173,28 @@ complete.RPolarsDataFrame <- function(
     out <- bind_rows_polars(already_exist, dont_already_exist)
     if (isTRUE(mo)) {
       if (isTRUE(is_grouped)) {
-        out <- out$sort(c(grps, vars))
+        out <- out$sort(c(grps, all_dots))
       } else {
-        out <- out$sort(vars)
+        out <- out$sort(all_dots)
       }
     }
   } else {
     if (isTRUE(is_grouped)) {
       out <- chain$join(
         data,
-        on = c(grps, vars),
-        how = "left",
-        join_nulls = TRUE
+        on = c(grps, all_dots),
+        how = "full",
+        join_nulls = TRUE,
+        coalesce = TRUE
       )
     } else {
-      out <- chain$join(data, on = vars, how = "left", join_nulls = TRUE)
+      out <- chain$join(
+        data,
+        on = all_dots,
+        how = "full",
+        join_nulls = TRUE,
+        coalesce = TRUE
+      )
     }
 
     if (length(fill) > 0) {
@@ -140,14 +202,13 @@ complete.RPolarsDataFrame <- function(
     }
   }
 
-  out <- if (is_grouped) {
-    out |>
+  out <- relocate(out, all_of(all_dots), .before = 1)
+
+  if (is_grouped) {
+    out <- out |>
       relocate(all_of(grps), .before = 1) |>
       group_by(all_of(grps), maintain_order = mo)
-  } else {
-    out
   }
-
   add_tidypolars_class(out)
 }