+++ tests/testthat/test-across-lazy.R
@@ -73,7 +73,7 @@ test_that("anonymous functions has to return a Polars expression", {
       gear_mean = mean(gear),
       gear_std = sd(gear),
       carb_mean = mean(carb),
-      carb_std = sd(carb),
+      carb_std = sd(carb)
     )
   )
 
@@ -102,7 +102,7 @@ test_that("anonymous functions has to return a Polars expression", {
   expect_snapshot_lazy(
     mutate(
       test,
-      across(.cols = contains("a"), \(x) mean(x)),
+      across(.cols = contains("a"), \(x) mean(x))
     ),
     error = TRUE
   )