+++ tests/testthat/test-bind_cols-lazy.R
@@ -6,11 +6,11 @@ test_that("returns custom class", {
   l <- list(
     polars::pl$LazyFrame(
       x = sample(letters, 20),
-      y = sample(1:100, 20)
+      y = sample.int(100, 20)
     ),
     polars::pl$LazyFrame(
       a = sample(letters, 20),
-      z = sample(1:100, 20)
+      z = sample.int(100, 20)
     )
   )
 
@@ -21,11 +21,11 @@ test_that("basic behavior with list works", {
   l <- list(
     polars::pl$LazyFrame(
       x = sample(letters, 20),
-      y = sample(1:100, 20)
+      y = sample.int(100, 20)
     ),
     polars::pl$LazyFrame(
       a = sample(letters, 20),
-      z = sample(1:100, 20)
+      z = sample.int(100, 20)
     )
   )
   expect_dim(
@@ -37,11 +37,11 @@ test_that("basic behavior with list works", {
 test_that("passing individual elements works", {
   p1 <- pl$LazyFrame(
     x = sample(letters, 20),
-    y = sample(1:100, 20)
+    y = sample.int(100, 20)
   )
   p2 <- pl$LazyFrame(
     z = sample(letters, 20),
-    w = sample(1:100, 20)
+    w = sample.int(100, 20)
   )
 
   expect_equal_lazy(
@@ -54,11 +54,11 @@ test_that("error if not all elements don't have the same class", {
   l <- list(
     data.frame(
       x = sample(letters, 20),
-      y = sample(1:100, 20)
+      y = sample.int(100, 20)
     ),
     polars::pl$LazyFrame(
       y = sample(letters, 20),
-      z = sample(1:100, 20)
+      z = sample.int(100, 20)
     )
   )
 
@@ -73,11 +73,11 @@ test_that("elements must be either all DataFrames or all LazyFrames", {
   l <- list(
     polars::pl$LazyFrame(
       x = sample(letters, 20),
-      y = sample(1:100, 20)
+      y = sample.int(100, 20)
     ),
     polars::pl$LazyFrame(
       y = sample(letters, 20),
-      z = sample(1:100, 20)
+      z = sample.int(100, 20)
     )
   )
 
@@ -91,15 +91,15 @@ test_that("can only bind more than 2 elements if DataFrame", {
   l <- list(
     polars::pl$LazyFrame(
       x = sample(letters, 20),
-      y = sample(1:100, 20)
+      y = sample.int(100, 20)
     ),
     polars::pl$LazyFrame(
       a = sample(letters, 20),
-      z = sample(1:100, 20)
+      z = sample.int(100, 20)
     ),
     polars::pl$LazyFrame(
       v = sample(letters, 20),
-      w = sample(1:100, 20)
+      w = sample.int(100, 20)
     )
   )
 
@@ -127,13 +127,13 @@ test_that("arg .name_repair works", {
     pl$LazyFrame(z = 1, x = 2, y = 3)
   )
 
-  expect_equal_lazy(
-    names(bind_cols_polars(l)),
+  expect_named(
+    bind_cols_polars(l),
     c("a", "x...2", "y...3", "z", "x...5", "y...6")
   )
 
-  expect_equal_lazy(
-    names(bind_cols_polars(l, .name_repair = "universal")),
+  expect_named(
+    bind_cols_polars(l, .name_repair = "universal"),
     c("a", "x...2", "y...3", "z", "x...5", "y...6")
   )
 
@@ -160,13 +160,10 @@ test_that("arg .name_repair works", {
     pl$LazyFrame(x = 1)$rename(list(x = " "))
   )
 
-  expect_equal_lazy(
-    names(bind_cols_polars(l)),
-    c(" ...1", " ...2")
-  )
+  expect_named(bind_cols_polars(l), c(" ...1", " ...2"))
 
-  expect_equal_lazy(
-    names(bind_cols_polars(l, .name_repair = "universal")),
+  expect_named(
+    bind_cols_polars(l, .name_repair = "universal"),
     c("....1", "....2")
   )
 })