+++ tests/testthat/test-count-lazy.R
@@ -94,15 +94,9 @@ test_that("add_count works on grouped data", {
     "am"
   )
 
-  expect_equal_lazy(
-    attr(count(test_grp), "maintain_grp_order"),
-    TRUE
-  )
+  expect_true(attr(count(test_grp), "maintain_grp_order"))
 
-  expect_equal_lazy(
-    attr(add_count(test_grp), "maintain_grp_order"),
-    TRUE
-  )
+  expect_true(attr(add_count(test_grp), "maintain_grp_order"))
 })
 
 test_that("message if overwriting variable", {