+++ tests/testthat/test-fill.R
@@ -76,9 +76,8 @@ test_that("works with grouped data", {
     "grp"
   )
 
-  expect_equal(
+  expect_true(
     fill(pl_grouped, everything(), .direction = "down") |>
-      attr("maintain_grp_order"),
-    TRUE
+      attr("maintain_grp_order")
   )
 })