+++ NEWS.md
@@ -15,10 +15,20 @@
 
 * Add support for argument `explicit` in `tidyr::complete()`.
 
+* Add option to keep track of filenames in `scan_csv_polars()` (#171, @ginolhac).
+
+* Add partial support for `seq()` (argument `length.out` is not supported) and 
+  `seq_len()`.
+
+* `complete()` now accepts named elements, e.g. `complete(df, group, value = 1:4)`
+  (#176).
+
 ## Bug fixes
 
 * Fix edge cases in the `tidypolars` implementation of `stringr::str_sub()`
   and `substr()` compared to their original implementation (#159). 
+
+* `arrange()` now places `NA` values last, like `dplyr`.
 
 # tidypolars 0.12.0